restructured videos components
