+++ frontend/app/components/navigator/style.ts
