+++ frontend/app/components/videoContent/VideoContent.tsx
@@ -1,11 +1,15 @@
 "use client";
 import Grid from '@mui/material/Grid';
 import { Box, CssBaseline, ThemeProvider, Card } from "@mui/material";
-import VideoPlayer from '../videoPlayer/VideoPlayer';
+import VideoPlayer, { VideoPlayerProvider } from '../videoPlayer/VideoPlayer';
 import { useState, useRef, createContext } from 'react';
 import VideoNavigator from '../navigator/VideoNavigator';
-import { IMediaHistory } from '../interfaces/interfaces';
-import darkTheme from '../theme';
+import { IMediaHistory } from '../../interfaces/interfaces';
+import darkTheme from '../../theme';
+import ControllerBar from '../controllerBar/ControllerBar';
+
+const MAX_MEDIA_COUNT = 25;
+
 
 export const MediaIndexContext = createContext<{
   mediaIndex: number,
@@ -18,22 +22,25 @@ export const MediaHistoryContext = createContext<{
     setMediaHistory: 
     React.Dispatch<React.SetStateAction<IMediaHistory>>
   }>(null!);
+
+export const ProgressContext = createContext<{
+  progress: number,
+    setProgress: React.Dispatch<React.SetStateAction<number>>,
+  }>(null!);
   
-export const MaxMediaIndexContext = createContext<number>(null!);
+export const MaxMediaCountContext = createContext<number>(null!);
 
 export const getRandomInt = (min: number, max: number): number => (
     Math.floor(Math.random() * (max - min)) + min
 )
 
 export default function VideoContent() { 
-  const maxMediaIndex = 25;
-  const [mediaIndex, setMediaIndex] = useState(getRandomInt(0, maxMediaIndex));
+  const [mediaIndex, setMediaIndex] = useState(getRandomInt(0, MAX_MEDIA_COUNT));
   const apiUrl = useRef(`http://localhost:8000/api/media`);
   const [mediaHistory, setMediaHistory] = useState<IMediaHistory>({
     currIndex: -1,
     mediaHistory: []
   });
-
 
 
   return (
@@ -42,8 +49,9 @@ export default function VideoContent() {
       <main>
         <Box maxWidth="xl" margin={2}>    
           <MediaHistoryContext.Provider value={{ mediaHistory, setMediaHistory }}>  
-            <MaxMediaIndexContext.Provider value={maxMediaIndex}>
+            <MaxMediaCountContext.Provider value={MAX_MEDIA_COUNT}>
               <MediaIndexContext.Provider value={{ mediaIndex, setMediaIndex }}>
+                <VideoPlayerProvider>
                   <Card>
                     <Grid container justifyContent="flex-start" alignItems="center">
                       <Grid size="auto" paddingLeft={2} paddingRight={2}>
@@ -53,9 +61,15 @@ export default function VideoContent() {
                         <VideoPlayer title="the title" src={apiUrl.current + `?index=${mediaIndex}`} />            
                       </Grid>
                     </Grid>
+                    <Grid container justifyContent="flex-start" alignItems="center">
+                      <Grid size={12}>
+                          <ControllerBar />
+                      </Grid>
+                    </Grid>
                   </Card>
+                </VideoPlayerProvider>
               </MediaIndexContext.Provider>
-            </MaxMediaIndexContext.Provider>
+            </MaxMediaCountContext.Provider>
           </MediaHistoryContext.Provider>
         </Box>
       </main>
