+++ frontend/app/components/videoPlayer/styles.ts
@@ -23,26 +23,4 @@ export const VideoFlexContainer = styled(Box)(() => ({
   position: 'relative',
 }));
 
-export const SliderBox = styled(Box)(() => ({
-  display: 'flex', 
-  alignItems: 'center', 
-  justifyContent: 'flex-start', 
-  width: '90%'
-}));
-
-export const VolumeBox = styled(Box)(() => ({
-  display: 'flex', 
-  alignItems: 'center', 
-  justifyContent: 'center',
-  width: '10%'
-}));
-
-export const VolumeSliderContainer = styled('div')(() => ({
-  position: 'absolute',
-  bottom: '3.5rem',
-  left: '50%',
-  transform: 'translateX(-50%)',
-  display: 'none',
-  zIndex: 2,
-}));
